* Added SAFE_RELEASE macro to PlatformSetup, I see it enough to warrant that I think... * BUGFIX: Fixed possible problem with extra letters being added when multiple input components were used at once * Added gamepad support classes * (Win) Added directx gamepad support for Windows - no public examples use it yet * (iOS) Added gamepad support for the 60beat iOS gamepad controller - no public examples use it yet, but I have some docs in GamepadManager.h * "PLATFORM_IOS" is now defined in the PlatformSetup for IOS projects.  The plan is to add PLATFORM_WEBOS, and so forth to keep things consistent * "PLATFORM_WINDOWS" is now defined in the PlatformSetup for windows projects * (IOS/Denshion) - Tweak so we we wait for Denshion to start up instead of letting it be synchronous with other events, this is primarily to get around issues of the 60beat gamepad on iOS getting confused when initting at the same time * Added new alignment case: ALIGNMENT_LEFT_CENTER.  I'm really not happy with the naming of these but.. * Added BaseApp:SetAllowScreenDimming(bool) - only respected by iOS at the moment * (possibly breaking change) Enums VIRTUAL_KEY_RELEASE and VIRTUAL_KEY_PRESS were reversed, to be compatible with the idea of "1" being pressed, instead of 0, which seemed weird

+++ proton/shared/util/TextScanner.cpp
@@ -135,7 +135,7 @@ string TextScanner::GetAll()
 
 std::string TextScanner::GetParmStringFromLine( int lineNum, int index, string token /*= "|"*/ )
 {
-	assert(lineNum >= 0 && lineNum < m_lines.size());
+	assert(lineNum >= 0 && (uint32)lineNum < m_lines.size());
 	assert(token.size() == 1 && "We don't actually support a non char delim yet");
 	return SeparateStringSTL(m_lines[lineNum], index, token[0]);
 }
